requirements.txt
python-telegram-bot==20.7
pyautogui==0.9.54
Pillow==10.0.0
python-dotenv==1.0.0
pip-system-certs
pyperclip
loguru
pywinauto; sys_platform == "win32"
